betting_arbitrage_bot/requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
python-telegram-bot>=20.0
streamlit>=1.30.0
pandas>=2.1.0
pydantic-settings>=2.1.0
loguru>=0.7.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
